+++ gatsby-theme-lannonbr/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0 - June 22, 2019
+
+Feat: Updated navbar to be toggleable fullscreen component on mobile layouts
+Feat: Switched dark mode toggle from checkbox w/ label to a button with a SVG within.
+
 # 1.1.0 - May 17, 2019
 
 Updating Dependencies and removing `gatsby-plugin-compile-es6-packages` as it is built into Gatsby now.